# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: NOMOS API
  version: 0.1.0
  description: |
    에이전트 협업 조율 서버. 수동 테스트 절차는 `docs/manual-test.md`를 함께 보라.

    **토큰이 두 종류다.** 오른쪽 위 Authorize를 누르면 칸이 두 개 뜬다.

    | 칸 | 어떤 토큰 | 받는 법 |
    |---|---|---|
    | `userToken` | 사람 토큰 (`kind: user`) | `POST /auth/login` — 시드 계정 `rep` / `manual-test-1234` |
    | `agentToken` | 에이전트 토큰 (`kind: agent`) | `npm run seed` 출력의 **be-laptop access token**, 또는 `POST /agents/token/refresh` |

    서로의 API를 부를 수 없다. 사람 토큰으로 태스크 API를 부르면 401이다.

    access token은 **1시간**이다. 만료되면 `POST /agents/token/refresh`로 다시 받는다.
servers:
  - url: http://localhost:3000/api
    description: 로컬 개발 서버
components:
  securitySchemes:
    userToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 사람 토큰. `POST /auth/login`의 `data.accessToken`을 붙여넣는다.
    agentToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 에이전트 토큰. 시드 출력의 access token 또는 refresh 응답의 값을 붙여넣는다.
...
  parameters:
...
    OrgId:
      name: orgId
      in: path
      required: true
      schema: { type: string, format: uuid }
      example: '{{ORG_ID}}'
...
paths:
...
  /orgs/{orgId}/repos:
    post:
      tags: [레포·경로]
      summary: 레포 연결 (조직 멤버)
      description: |
        연결하면 기본 경로 규칙 **15개**가 함께 생성된다(`source=seed`, `owner_role`은 전부 NULL).

        **대표 전용이 아니다.** `repos` 행은 그 자체로 권한을 만들지 않는다 — 판정에 쓰이는 경로 규칙은
        `project_repos`에 조인된 레포만 읽으므로(`listProjectRepoPaths`), 연결만 된 레포는 어떤
        에이전트도 건드릴 수 없다. 실제 관문인 소유권 지정(`PATCH .../paths/{pathId}`)과
        프로젝트 투입(`POST /orgs/{orgId}/projects`의 `repoIds`)은 대표 전용으로 남는다.
      security: [{ userToken: [] }]
      parameters: [{ $ref: '#/components/parameters/OrgId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repos]
              properties:
                repos:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [fullName]
                    properties:
                      fullName: { type: string }
                      githubRepoId: { type: integer }
                      defaultBranch: { type: string }
            example:
              repos: [{ fullName: acme/extra-repo }]
      responses:
        '201':
          description: 연결됨
          content:
            application/json:
              example:
                data: { repos: [{ id: 2ea71bc6-05b8-4725-9fa0-2b2eec9d3674, fullName: acme/extra-repo, seededPathCount: 15 }] }
        '409':
          description: 이미 연결된 레포
          content:
            application/json:
              example:
                error: { code: REPO_ALREADY_CONNECTED, message: repository is already connected }
  /repos/{repoId}:
    patch:
      tags: [레포·경로]
      summary: V3가 커밋을 읽을 곳 수정 (대표 전용)
      description: |
        `githubRepoId`(운영 github 모드 — 이름이 아니라 id로 찾는다)와 `cloneUrl`(로컬 mirror 모드)만 바꾼다. 넘긴 키만 바뀌고 `null`은 비운다.

        **`cloneUrl`은 서버가 `git clone`에 그대로 넘긴다.** 그래서 `https://`(자격 증명 없이), `file:///`, 로컬 절대 경로만 받는다.
        `-`로 시작(옵션 주입), `ext::`, `ssh://`·`git@host:`, `http://`, `user:pass@`는 400이고 사유가 메시지에 담긴다.

        같은 조직의 다른 레포가 이미 그 `githubRepoId`를 가리키면 409. 경로 규칙이 아니므로 `policy_hash`는 바뀌지 않는다.
      security: [{ userToken: [] }]
      parameters:
        - name: repoId
          in: path
          required: true
          schema: { type: string, format: uuid }
          example: '{{REPO_API}}'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                githubRepoId: { type: integer, nullable: true, minimum: 1 }
                cloneUrl: { type: string, nullable: true, maxLength: 2048 }
            example: { githubRepoId: 987654321 }
      responses:
        '200':
          description: 바뀐 설정
          content:
            application/json:
              example:
                data:
                  repo: { id: '{{REPO_API}}', fullName: acme/study-api, githubRepoId: 987654321, defaultBranch: main, devBranch: dev, cloneUrl: null }
        '400':
          description: 빈 바디, 또는 허용되지 않는 cloneUrl
          content:
            application/json:
              example:
                error: { code: VALIDATION_ERROR, message: 'clone_url 이 -로 시작한다(git 옵션으로 읽힌다) — https://host/owner/repo.git(자격 증명 없이) 또는 로컬 절대 경로만 받는다' }
        '403':
          description: 대표가 아니거나 다른 조직
          content:
            application/json:
              example:
                error: { code: NOT_REPRESENTATIVE, message: this action requires the organization representative }
        '409':
          description: 같은 GitHub 레포를 이미 가리키는 레포 행이 있음
          content:
            application/json:
              example:
                error: { code: REPO_ALREADY_CONNECTED, message: 'github_repo_id 987654321 is already used by acme/study-web (...)' }
...